Break the maker in to "sub-makers"
This is my first attempt at breaking up the reconciliation into more
maintainable pieces that follow a particular pattern.  This is very
much a work in progress and I expect the abstractions to be improved or
totally refactored as time goes on.

The idea is that there is a sub-maker for each section of the base
config.  Each sub-maker has two responsibilities:

  1. Reconcilie the dependent resources for the target app
  2. Contribute its relevant section to the app's JSON config

Both responsibilites need to take into account which providers are
configured for each section of base config.