Trying to fix kafka listener tests
I ended up doing a "get" rather than a "list" in the maker.

I also added a kafka resource to the test suite, but for some reason the
listeners are not getting persisted.  you can see in the test that I
create the resource and then immediately get it and see if the listeners
are there.  I print the info, and it is not there!