Change ClowdObject so it does not override runtime.Object
GetName and GetNamespace cannot be overridden by ClowdObject because it
affects how the k8s client fetches data.  Therefore I just stuck the
word "Clowd" in the function names to make GetClowdName and
GetClowdNamespace.  This is probably gross, but I'm not sure if there is
a better way, and I just wanted to get it working for now.