Add app-interface provider for kafka
This is a very basic provider that assumes both Kafka and the topics are
managed by app-interface.  It assumes that app-interface has already
applied the correct topic resources and always sets the topic name to be
the same as the requested topic name.

It would be reasonable in the future to add some validation to ensure
that all requested topics have been applied by app-interface, though it
may not be worth the effort if we remove this responsiblity from
app-interface in the short term.