Overwrite port number for liveness/readiness probes
Devs currently have to manually copy the port number from the
environment spec to their liveness/readiness probes in their app spec.
This would automatically update the probes with the correct port number
given an app uses the HTTPGet handler.

The allowed values are "public", "private", "metrics", and the
corresponding (correct) port numbers.  All numbers are updated to the
string value in the spec.

Any other value is defaulted to "public" if the public port is enabled,
or "metrics" if not.

The ports listed in the container specs now include both the "private"
port, and the "web" port has been renamed to "public".