Add prometheus hostname to clowdenv status field
This will allow an entity to look at clowdenv.status.prometheus.hostname
to know how to connect to the Prometheus instance associated with this
environment.

When in app-interface mode, the hostname will just be copied over from
clowdenv.spec.providers.metrics.prometheus.appInterfaceHostname. When in
operator mode, the hostname will be generated based on the target
namespace of the clowdenv.

More fields may be added to clowdenv.status.prometheus down the road,
e.g. TLS, credentials, etc.