Calc authsidecar-confighash w/ only secret data
The current calculation returns a different result the second time the
hash is calculated, which results in a needless deployment update (and
rolling deployment).

Note that I was unable to reproduce locally to verify what fields cause
the hash to change, but I believe it must be fields other than data.

If someone can reproduce in an environment with sufficient logging it
may be worth adding the following line to determine what fields
specifically caused problems:

```
web.Log.Info("AuthConfigJson", "json", string(jsonData), "hash", hash)
```

Note this actually makes it so that `authsidecar-confighash` works more
like the regular `confighash`, hashing only the data, and not the cached
secret.