Cert auth (#851)
* Added cert auth

This commit adds cert auth to Clowder for single namespace ephemeral envs. It works by adding another gateway
pod that has cert auth capability. The certificate needs to be registered using one of the traditional endpoints.
Details are in the documentation. There were a few spec additions in this PR, but no removals/changes.