fix: align cloudwatch provider with documentation
We observed the value for the `Type` field is an empty
string when we are using the cloudwatch provider.

We think this could fix the root cause by setting
the value for the clouderwatch logging provider, and this
value would be propagated to the configuration injected
in the final deployed pods.

See: https://github.com/RedHatInsights/clowder/blob/master/docsmd/providers/logging.md

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>