Adds the `user_id` claim to Keycloak JWTs
This adds the missing `user_id` claim for the `cloud-services` client in Keycloak
for ephemeral. Since this value is required in the `x-rh-identity` schema [1], some
services will expect/require this to be populated on the identity. When it's not
an attribute on the client/users, it will not be included in the JWT and thus
cause issues for upstream services relying on this.

This change ensures that it exists on the token for the `jdoe` user, meaning the
gateway (will require an update to crcauthlib) will pass the value along on the
identity header [2].

[1] https://github.com/RedHatInsights/identity-schemas/blob/main/3scale/schema.json#L103
[2] https://github.com/RedHatInsights/crcauthlib/blob/master/crcauthlib.go#L30